Tests now pass but need lots of enhancing
